fix: Bug in my tests; I was overlapping ranges of values.
This was a classic copy/paste error.  The tests were designed for
integers which required fewer than 8 bytes of storage, but I tried
to use them on integers that require 16 bytes of storage without
modifying the offsets correctly.  As a result, one value's byte
range overlapped another value's byte range.  The result is exactly
what you'd expect; chaos.  Now fixed so that the ranges are widely
separated.