fix: Added #![allow(clippy::float_cmp)].
Clippy doesn't like it when floats are compared directly as floating
point arithmetic is imprecise.  Normally, this is a good idea, but
it is a bad idea for binary-layout's tests, which need to come back
EXACTLY the same, or there's an error somewhere.  I've turned off
this lint for the tests.