Add f32/f64 support (#8)
* WIP: Added support for i128/u128, but currently failing tests.

* fix: Bug in my tests; I was overlapping ranges of values.

This was a classic copy/paste error.  The tests were designed for
integers which required fewer than 8 bytes of storage, but I tried
to use them on integers that require 16 bytes of storage without
modifying the offsets correctly.  As a result, one value's byte
range overlapped another value's byte range.  The result is exactly
what you'd expect; chaos.  Now fixed so that the ranges are widely
separated.

* test: Added tests for u128/i128 to fields::wrapped.

* feat: Updated ChangeLog and doc comments to mention support for i128 and u128.

* fix: Forgot to run gen_readme.md.

* feat: Added f32 and f64 support.

This is mostly just copy/paste/modify of int_field!(), modifying things
enough to support floats.  The rest is copy/paste/modifying tests so that
they correctly deal with the new f32/f64 types.

* fix: Added #![allow(clippy::float_cmp)].

Clippy doesn't like it when floats are compared directly as floating
point arithmetic is imprecise.  Normally, this is a good idea, but
it is a bad idea for binary-layout's tests, which need to come back
EXACTLY the same, or there's an error somewhere.  I've turned off
this lint for the tests.