Ran `preen_rust.sh` across everything (#12)
* feat: Added support for the unit type ('()').

* fix: Silenced some clippy warnings.

I implemented the FieldCopyAccess trait on PrimitiveField for the unit
type (`()`).  I want to make it BLINDINGLY OBVIOUS that this is what
I'm doing, which is why I want to keep the unit there.  So I'm
silencing this warning.

* fix: Even more clippy silencing.

Once again, I'm silencing warnings about comparing unit with other
objects.  I *really want* to be doing this comparison, if for no
other reason than it may one day fail.

* fix: Added a not in the change log about supporting unit (`()`) now.

* test: Added the test_layout_with_unit() test to verify that define_layout!() works with unit.

@smessmer pointed out that the macro could have issues dealing with
unit in the field definitions.  This adds a small set of tests that
test out how define_layout!() handles unit in different field positions,
and in combination with different types.

* docs: Moved the entry regarding zero-sized type support lower down in the docs.

As @smessmer pointed out, it's unlikely that anyone is going to actually
need this kind of support, it's more for helping derive macros that may
not be able to handle zero-sized types that don't have this support.

* style: Ran preen_rust.sh across everything.